+++ src/ts/platforms/apple-appstore/appstore-adapter.ts
@@ -219,10 +219,12 @@ namespace CdvPurchase {
             /** Notify the store that the receipts have been updated */
             private _receiptsUpdated() {
                 if (this._receipt) {
+                    this.log.debug("receipt updated and ready.");
                     this.context.listener.receiptsUpdated(Platform.APPLE_APPSTORE, [this._receipt, this.pseudoReceipt]);
                     this.context.listener.receiptsReady(Platform.APPLE_APPSTORE);
                 }
                 else {
+                    this.log.debug("receipt updated.");
                     this.context.listener.receiptsUpdated(Platform.APPLE_APPSTORE, [this.pseudoReceipt]);
                 }
             }
@@ -340,13 +342,15 @@ namespace CdvPurchase {
                 });
             }
 
+            supportsParallelLoading = true;
+
             loadReceipts(): Promise<Receipt[]> {
                 return new Promise((resolve) => {
                     setTimeout(() => {
                         this.initializeAppReceipt(() => {
                             this.receiptsUpdated();
                             if (this._receipt) {
-                               resolve([this._receipt, this.pseudoReceipt]);
+                                resolve([this._receipt, this.pseudoReceipt]);
                             }
                             else {
                                 resolve([this.pseudoReceipt]);
@@ -467,7 +471,7 @@ namespace CdvPurchase {
                             discountType: discount.type,
                         });
                     });
-                    if (!valid.discounts && valid.introPrice) {
+                    if ((valid.discounts?.length ?? 0) === 0 && valid.introPrice) {
                         // sometime apple returns the discounts in the deprecated "introductory" info
                         // we create a special "discount" with the id "intro" to check for eligibility.
                         eligibilityRequests.push({
@@ -530,7 +534,7 @@ namespace CdvPurchase {
                                         product?.refresh(valid, this.context.apiDecorators, eligibilities);
                                     }
                                     else {
-                                        this.log.debug('registering existing product');
+                                        this.log.debug('registering new product');
                                         product = new SKProduct(valid, p, this.context.apiDecorators, eligibilities);
                                         this._products.push(product);
                                     }